requirements.txt
openai
tiktoken
streamlit
python-dotenv